модели добавление связей один ко многим 0.95

+++ src/db/models/departments.js
@@ -22,7 +22,7 @@ module.exports = (sequelize, DataTypes) => {
     },
 
     sRegionId: {
-      type: DataTypes.INT(11).UNSIGNED,
+      type: DataTypes.INTEGER(11).UNSIGNED,
       allowNull: true,
       defaultValue: null
     },
