+++ src/db/models/sMaterialy.js
@@ -22,13 +22,13 @@ module.exports = (sequelize, DataTypes) => {
     },
 
     sMaterialTypeId: {
-      type: DataTypes.INT(11).UNSIGNED,
+      type: DataTypes.INTEGER(11).UNSIGNED,
       allowNull: true,
       defaultValue: null
     },
 
     sMaterialAnimalTypeId: {
-      type: DataTypes.INT(11).UNSIGNED,
+      type: DataTypes.INTEGER(11).UNSIGNED,
       allowNull: true,
       defaultValue: null
     },
